requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
dill
flask
mysql.connector
#-e.